requirements.txt
# ===============================
# ZoL0-master requirements.txt
# ===============================

# --- Core scientific stack ---
numpy==1.26.4
pandas==2.2.2
scipy==1.13.0
scikit-learn==1.4.2
joblib==1.4.2

# --- Deep learning & transformers ---
torch==2.7.0
tensorflow==2.16.2
keras==3.10.0              # TF ≥2.16 wymaga Keras 3
xgboost==2.0.3
transformers==4.40.2
tokenizers==0.19.1         # zakres <0.20, zgodny z transformers 4.40.x
onnx==1.17.0               # ostatnia wersja na protobuf 3.x
onnxruntime==1.22.0
tf2onnx==1.16.1
protobuf==3.20.3           # wspólna baza dla onnx/ort/tf2onnx
flatbuffers==24.3.25       # brakowałoby tf2onnx
six==1.16.0                # j.w.

skl2onnx==1.18.0
ta==0.10.3                 # czysty Python – zamiennik TA-Lib

# --- RL / symulacje ---
gymnasium==0.29.1

# --- Web API & micro-services ---
fastapi==0.111.0
pydantic==2.7.1
uvicorn[standard]==0.30.0
flask==3.0.3
aiohttp==3.9.5
websocket-client==1.8.0
pybit==5.11.0

# --- Data / I-O / utils ---
sqlalchemy==2.0.30
pyyaml==6.0.1
requests==2.31.0
dateparser==1.2.0
ipywidgets==8.1.2
numba==0.59.1
psutil==5.9.8
watchdog==3.0.0
astunparse==1.6.3
typing_extensions==4.12.0
python-dotenv==1.0.1
polars

# --- Crypto / security ---
cryptography==42.0.5
pyjwt==2.8.0

# --- Visualisation ---
matplotlib==3.8.4
mplfinance==0.12.10b0

# --- Dev / CI ---
pytest==8.2.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
coverage==7.5.1
black==24.4.2
isort==5.13.2
flake8==7.0.0
pipdeptree==2.15.1
TA-Lib==0.4.28
